+++ src/aunit/Test.h
@@ -57,10 +57,31 @@ void test_ ## name :: once()
 test_ ## name :: test_ ## name() : Test(F(#name)) {}\
 void test_ ## name :: loop()
 
+// Create an extern reference to a test() test case object defined elsewhere.
+// This is only necessary if you use assertTestXxx() or checkTestXxx() when the
+// test is in another file (or defined after the assertion on it).
+#define externTest(name) struct test_ ## name : aunit::TestOnce {\
+  test_ ## name();\
+  void once();\
+};\
+extern test_##name test_##name##_instance
+
+// Create an extern reference to a testing() test case object defined
+// elsewhere.  This is only necessary if you use assertTestXxx() or
+// checkTestXxx() when the test is in another file (or defined after the
+// assertion on it).
+#define externTesting(name) struct test_ ## name : aunit::Test {\
+  test_ ## name();\
+  void loop();\
+};\
+extern test_##name test_##name##_instance
+
 namespace aunit {
 
 class Test {
   public:
+    // Don't change the order of Passed, Failed, Skipped or Expired without
+    // looking at the isDone() method.
     static const uint8_t kStatusNew = 0;
     static const uint8_t kStatusSetup = 1;
     static const uint8_t kStatusPassed = 2;
@@ -120,6 +141,36 @@ class Test {
      * Also makes setNext() method unnecessary.
      */
     Test** getNext() { return &mNext; }
+
+    /** Return true if test is done (passed, failed, skipped, expired). */
+    bool isDone() { return mStatus >= kStatusPassed; }
+
+    /** Return true if test is done (passed, failed, skipped, expired). */
+    bool isNotDone() { return !isDone(); }
+
+    /** Return true if test is passed. */
+    bool isPassed() { return mStatus == kStatusPassed; }
+
+    /** Return true if test is passed. */
+    bool isNotPassed() { return !isPassed(); }
+
+    /** Return true if test is failed. */
+    bool isFailed() { return mStatus == kStatusFailed; }
+
+    /** Return true if test is failed. */
+    bool isNotFailed() { return !isFailed(); }
+
+    /** Return true if test isNot skipped. */
+    bool isSkipped() { return mStatus == kStatusSkipped; }
+
+    /** Return true if test isNot skipped. */
+    bool isNotSkipped() { return !isSkipped(); }
+
+    /** Return true if test is expired. */
+    bool isExpired() { return mStatus == kStatusExpired; }
+
+    /** Return true if test is expired. */
+    bool isNotExpired() { return !isExpired(); }
 
     /** Mark the test as skipped. */
     void skip() { mStatus = kStatusSkipped; }